+++ client/src/App.tsx
@@ -5,6 +5,7 @@ import Main from './features/main';
 import Login from './features/login';
 import Write from './features/write';
 import { RouteCustoms } from './modules/RouteCustoms';
+import Archive from './features/archive';
 
 
 const App = () => {
@@ -15,6 +16,7 @@ const App = () => {
         <Route path='*' element={<></>} />
         <Route path='/login' element={<Login />} />
         <Route path="/report/*" element={<Write />} />
+        <Route path="/archive/*" element={<Archive />} />
       </Routes>
     </Router>
   );