+++ client/src/features/archive/index.tsx
@@ -1,0 +1,14 @@
+import { Route, Routes } from 'react-router-dom';
+import Overview from './overview';
+
+export const Archive = () => {
+
+  return (
+    <Routes>
+      <Route path='*' element={<Overview/>} />
+      <Route path='/:archiveID' element={<></>} />
+    </Routes>
+  );
+};
+
+export default Archive;