+++ client/src/scripts/IndexedDB/Archive/findArchive/by/id.ts
@@ -1,0 +1,5 @@
+import { getArchive } from "../../getArchive";
+
+export const byID = getArchive;
+
+export default byID;