+++ client/src/scripts/IndexedDB/Archive/findArchive/by/index.ts
@@ -1,0 +1,9 @@
+import byID from "./id"
+import { byAuthor } from "./author"
+import { byAuthorID } from "./authorID"
+
+export {
+  byID as id,
+  byAuthor as author,
+  byAuthorID as authorID
+}