+++ client/src/scripts/IndexedDB/Archive/index.ts
@@ -1,0 +1,15 @@
+import GetIDB from "./getIDB"
+import { getAllArchive } from "./getAllArchive"
+import { getArchive } from "./getArchive"
+import { setArchive } from "./setArchive"
+import { deleteArchive } from "./deleteArchive"
+import * as find from "./findArchive"
+
+export {
+  GetIDB,
+  getAllArchive,
+  getArchive,
+  setArchive,
+  deleteArchive,
+  find
+}