+++ client/src/scripts/IndexedDB/index.ts
@@ -1,7 +1,9 @@
 import * as FilesDB from './Files';
 import * as ReportDB from './Reports';
+import * as ArchiveDB from './Archive';
 
 export {
   FilesDB,
   ReportDB,
+  ArchiveDB
 }